Extract BaseClientManager into its own file; reduce client-manager.ts to interfaces
